RF: Move specialized underdocumented helper to only code using it
And its tests too.
