+++ datalad/cmd.py
@@ -16,7 +16,6 @@ import subprocess
 import sys
 import logging
 import os
-import shutil
 import shlex
 import atexit
 import functools
@@ -30,7 +29,7 @@ from .support import path as op
 from .consts import GIT_SSH_COMMAND
 from .dochelpers import exc_str
 from .support.exceptions import CommandError
-from .support.protocol import NullProtocol, DryRunProtocol, \
+from .support.protocol import NullProtocol, \
     ExecutionTimeProtocol, ExecutionTimeExternalsProtocol
 from .utils import (
     on_windows, get_tempfile_kwargs, assure_unicode, assure_bytes,
@@ -674,31 +673,3 @@ class GitRunner(Runner):
         # All communication here will be returned as unicode
         # TODO: do that instead within the super's run!
         return assure_unicode(out), assure_unicode(err)
-
-
-# ####
-# Preserve from previous version
-# TODO: document intention
-# ####
-# this one might get under Runner for better output/control
-def link_file_load(src, dst, dry_run=False):
-    """Just a little helper to hardlink files's load
-    """
-    dst_dir = op.dirname(dst)
-    if not op.exists(dst_dir):
-        os.makedirs(dst_dir)
-    if op.lexists(dst):
-        lgr.log(9, "Destination file %(dst)s exists. Removing it first", locals())
-        # TODO: how would it interact with git/git-annex
-        unlink(dst)
-    lgr.log(9, "Hardlinking %(src)s under %(dst)s", locals())
-    src_realpath = op.realpath(src)
-
-    try:
-        os.link(src_realpath, dst)
-    except AttributeError as e:
-        lgr.warn("Linking of %s failed (%s), copying file" % (src, e))
-        shutil.copyfile(src_realpath, dst)
-        shutil.copystat(src_realpath, dst)
-    else:
-        lgr.log(2, "Hardlinking finished")
